fix: fixed fork tests
